Fix the compliment doc completion code
There were a couple of problems with the doc function taken from
compliment. The main one was that compliment passes a Namespace object as
second argument and we were not handling that. Another problem was that both
analysis/sanitize-ns and the main code path were not handling the empty map
case properly.