Handle the user namespace for doc completions
It seems like compliment is passing the 'user namespace to the doc function by
default if the namespace is missing. This patch makes so that we check in both
cljs.user and cljs.core (in this order) if that is the case.
We check in cljs.core as well because that is were the doc for built-ins lives,
rather than in cljs.user.